feat: create new file moduelEdit (seperation of concerns) Add Module now working
